+++ engine/source/runtime/function/render/render_pass_base.h
@@ -25,6 +25,7 @@ namespace Piccolo
         virtual void setCommonInfo(RenderPassCommonInfo common_info);
         virtual void preparePassData(std::shared_ptr<RenderResourceBase> render_resource);
         virtual void initializeUIRenderBackend(WindowUI* window_ui);
+        virtual void shutdownUIRenderBackend();
 
     protected:
         std::shared_ptr<RHI>                m_rhi;