+++ engine/source/runtime/function/render/render_pipeline_base.h
@@ -34,6 +34,7 @@ namespace Piccolo
         virtual void deferredRender(std::shared_ptr<RHI> rhi, std::shared_ptr<RenderResourceBase> render_resource);
 
         void             initializeUIRenderBackend(WindowUI* window_ui);
+        void             shutdownUIRenderBackend();
         virtual uint32_t getGuidOfPickedMesh(const Vector2& picked_uv) = 0;
 
     protected: